[breaking] rename /deployer to /deploy
