+++ server/api/deploy/prebuilt.ts
@@ -4,7 +4,7 @@ import { StatusCodes } from "http-status-codes";
 import { walletAuthSchema } from "../../../core/schema";
 import { queueTx } from "../../../src/db/transactions/queueTx";
 import {
-  publishedDeployParamSchema,
+  prebuiltDeployParamSchema,
   standardResponseSchema,
 } from "../../helpers/sharedApiSchemas";
 import { txOverridesForWriteRequest } from "../../schemas/web3api-overrides";
@@ -12,10 +12,11 @@ import { getChainIdFromChain } from "../../utilities/chain";
 import { getSdk } from "../../utils/cache/getSdk";
 
 // INPUTS
-const requestSchema = publishedDeployParamSchema;
+const requestSchema = prebuiltDeployParamSchema;
 const requestBodySchema = Type.Object({
-  constructorParams: Type.Array(Type.Any(), {
-    description: "Constructor arguments for the deployment.",
+  // TODO need to type this
+  contractMetadata: Type.Any({
+    description: "Arguments for the deployment.",
   }),
   version: Type.Optional(
     Type.String({
@@ -28,7 +29,15 @@ const requestBodySchema = Type.Object({
 // Example for the Request Body
 requestBodySchema.examples = [
   {
-    constructorParams: ["0x1946267d81Fb8aDeeEa28e6B98bcD446c8248473"],
+    contractMetadata: {
+      name: `My Contract`,
+      description: "Contract deployed from web3 api",
+      primary_sale_recipient: "0x1946267d81Fb8aDeeEa28e6B98bcD446c8248473",
+      seller_fee_basis_points: 500,
+      fee_recipient: "0x1946267d81Fb8aDeeEa28e6B98bcD446c8248473",
+      platform_fee_basis_points: 10,
+      platform_fee_recipient: "0x1946267d81Fb8aDeeEa28e6B98bcD446c8248473",
+    },
   },
 ];
 
@@ -38,37 +47,38 @@ const responseSchema = Type.Object({
   deployedAddress: Type.Optional(Type.String()),
 });
 
-export async function deployPublished(fastify: FastifyInstance) {
+export async function deployPrebuilt(fastify: FastifyInstance) {
   fastify.route<{
     Params: Static<typeof requestSchema>;
     Reply: Static<typeof responseSchema>;
     Body: Static<typeof requestBodySchema>;
   }>({
     method: "POST",
-    url: "/deployer/:network/:publisher/:contract_name",
+    url: "/deploy/:network/:contract_type",
     schema: {
-      description: "Deploy published contract",
+      description: "Deploy prebuilt contract",
       tags: ["Deploy"],
-      operationId: "deployPublished",
+      operationId: "deployPrebuilt",
       params: requestSchema,
       body: requestBodySchema,
       headers: walletAuthSchema,
+      hide: true,
       response: {
         ...standardResponseSchema,
         [StatusCodes.OK]: responseSchema,
       },
     },
     handler: async (request, reply) => {
-      const { network, publisher, contract_name } = request.params;
-      const { constructorParams, version } = request.body;
+      const { network, contract_type } = request.params;
+      const { contractMetadata, version } = request.body;
       const chainId = getChainIdFromChain(network);
       const walletAddress = request.headers["x-wallet-address"] as string;
 
       const sdk = await getSdk({ chainId, walletAddress });
-      const tx = await sdk.deployer.deployReleasedContract.prepare(
-        publisher,
-        contract_name,
-        constructorParams,
+
+      const tx = await sdk.deployer.deployBuiltInContract.prepare(
+        contract_type,
+        contractMetadata,
         version,
       );
       const deployedAddress = await tx.simulate();
@@ -76,9 +86,7 @@ export async function deployPublished(fastify: FastifyInstance) {
       const queuedId = await queueTx({
         tx,
         chainId,
-        extension: "deploy-published",
-        deployedContractAddress: deployedAddress,
-        deployedContractType: contract_name,
+        extension: "deploy-prebuilt",
       });
       reply.status(StatusCodes.OK).send({
         deployedAddress,
