+++ server/api/deploy/prebuilts/multiwrap.ts
@@ -6,14 +6,12 @@ import { queueTx } from "../../../../src/db/transactions/queueTx";
 import { standardResponseSchema } from "../../../helpers/sharedApiSchemas";
 import {
   commonContractSchema,
-  commonPlatformFeeSchema,
-  commonPrimarySaleSchema,
   commonRoyaltySchema,
   commonSymbolSchema,
   commonTrustedForwarderSchema,
   prebuiltDeployContractParamSchema,
   prebuiltDeployResponseSchema,
-} from "../../../schemas/prebuilts/index";
+} from "../../../schemas/prebuilts";
 import { txOverridesForWriteRequest } from "../../../schemas/web3api-overrides";
 import { getChainIdFromChain } from "../../../utilities/chain";
 import { getSdk } from "../../../utils/cache/getSdk";
@@ -25,8 +23,6 @@ const requestBodySchema = Type.Object({
     ...commonContractSchema.properties,
     ...commonRoyaltySchema.properties,
     ...commonSymbolSchema.properties,
-    ...commonPlatformFeeSchema.properties,
-    ...commonPrimarySaleSchema.properties,
     ...commonTrustedForwarderSchema.properties,
   }),
   version: Type.Optional(
@@ -42,18 +38,18 @@ const requestBodySchema = Type.Object({
 // OUTPUT
 const responseSchema = prebuiltDeployResponseSchema;
 
-export async function deployPrebuiltEdition(fastify: FastifyInstance) {
+export async function deployPrebuiltMultiwrap(fastify: FastifyInstance) {
   fastify.route<{
     Params: Static<typeof requestSchema>;
     Reply: Static<typeof responseSchema>;
     Body: Static<typeof requestBodySchema>;
   }>({
     method: "POST",
-    url: "/deployer/:network/prebuilts/edition",
+    url: "/deploy/:network/prebuilts/multiwrap",
     schema: {
-      description: "Deploy prebuilt Edition contract",
+      description: "Deploy prebuilt Multiwrap contract",
       tags: ["Deploy"],
-      operationId: "deployPrebuiltEdition",
+      operationId: "deployPrebuiltMultiwrap",
       params: requestSchema,
       body: requestBodySchema,
       headers: walletAuthSchema,
@@ -69,9 +65,8 @@ export async function deployPrebuiltEdition(fastify: FastifyInstance) {
       const walletAddress = request.headers["x-wallet-address"] as string;
 
       const sdk = await getSdk({ chainId, walletAddress });
-
       const tx = await sdk.deployer.deployBuiltInContract.prepare(
-        "edition",
+        "multiwrap",
         contractMetadata,
         version,
       );
@@ -81,7 +76,7 @@ export async function deployPrebuiltEdition(fastify: FastifyInstance) {
         chainId,
         extension: "deploy-prebuilt",
         deployedContractAddress: deployedAddress,
-        deployedContractType: "edition",
+        deployedContractType: "multiwrap",
       });
       reply.status(StatusCodes.OK).send({
         result: {
