+++ server/api/deploy/prebuilts/split.ts
@@ -6,14 +6,10 @@ import { queueTx } from "../../../../src/db/transactions/queueTx";
 import { standardResponseSchema } from "../../../helpers/sharedApiSchemas";
 import {
   commonContractSchema,
-  commonPlatformFeeSchema,
-  commonPrimarySaleSchema,
-  commonRoyaltySchema,
-  commonSymbolSchema,
   commonTrustedForwarderSchema,
-  merkleSchema,
   prebuiltDeployContractParamSchema,
   prebuiltDeployResponseSchema,
+  splitRecipientInputSchema,
 } from "../../../schemas/prebuilts";
 import { txOverridesForWriteRequest } from "../../../schemas/web3api-overrides";
 import { getChainIdFromChain } from "../../../utilities/chain";
@@ -24,11 +20,7 @@ const requestSchema = prebuiltDeployContractParamSchema;
 const requestBodySchema = Type.Object({
   contractMetadata: Type.Object({
     ...commonContractSchema.properties,
-    ...commonRoyaltySchema.properties,
-    ...merkleSchema.properties,
-    ...commonSymbolSchema.properties,
-    ...commonPlatformFeeSchema.properties,
-    ...commonPrimarySaleSchema.properties,
+    recipients: Type.Array(splitRecipientInputSchema),
     ...commonTrustedForwarderSchema.properties,
   }),
   version: Type.Optional(
@@ -44,18 +36,18 @@ const requestBodySchema = Type.Object({
 // OUTPUT
 const responseSchema = prebuiltDeployResponseSchema;
 
-export async function deployPrebuiltEditionDrop(fastify: FastifyInstance) {
+export async function deployPrebuiltSplit(fastify: FastifyInstance) {
   fastify.route<{
     Params: Static<typeof requestSchema>;
     Reply: Static<typeof responseSchema>;
     Body: Static<typeof requestBodySchema>;
   }>({
     method: "POST",
-    url: "/deployer/:network/prebuilts/editionDrop",
+    url: "/deploy/:network/prebuilts/split",
     schema: {
-      description: "Deploy prebuilt Edition Drop contract",
+      description: "Deploy prebuilt Split contract",
       tags: ["Deploy"],
-      operationId: "deployPrebuiltEditionDrop",
+      operationId: "deployPrebuiltSplit",
       params: requestSchema,
       body: requestBodySchema,
       headers: walletAuthSchema,
@@ -65,6 +57,7 @@ export async function deployPrebuiltEditionDrop(fastify: FastifyInstance) {
       },
     },
     handler: async (request, reply) => {
+      //TODO add x-wallet-address to headers
       const { network } = request.params;
       const { contractMetadata, version } = request.body;
       const chainId = getChainIdFromChain(network);
@@ -72,19 +65,19 @@ export async function deployPrebuiltEditionDrop(fastify: FastifyInstance) {
 
       const sdk = await getSdk({ chainId, walletAddress });
       const tx = await sdk.deployer.deployBuiltInContract.prepare(
-        "edition-drop",
+        "split",
         contractMetadata,
         version,
       );
       const deployedAddress = await tx.simulate();
+
       const queuedId = await queueTx({
         tx,
         chainId,
         extension: "deploy-prebuilt",
         deployedContractAddress: deployedAddress,
-        deployedContractType: "edition-drop",
+        deployedContractType: "split",
       });
-
       reply.status(StatusCodes.OK).send({
         result: {
           deployedAddress,
