+++ test/e2e/contract/contract.test.ts
@@ -13,7 +13,7 @@ describe.skip("Deploy EditionDrop to Test : Contract Endpoints", () => {
   before(async () => {
     createdServerInstance = await createServer("Test-Suite");
     const contractDeployedResponse = await request(createdServerInstance.server)
-      .post("/deployer/localhost/prebuilts/editionDrop")
+      .post("/deploy/localhost/prebuilts/editionDrop")
       .set("Authorization", `Bearer ${env.THIRDWEB_API_SECRET_KEY}`)
       .set("x-wallet-address", "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266")
       .send({
